Adding strong names and dependency range for Json.Net
Adding strong names for all the assemblies apart from:
Nethereum.HDWallet as it uses NBitcoin and is not strongnamed.
Trezor and Ledger as both use Trezor.Net and Ledger.Net and are not strongnamed.
Net351 and Net461 AOT builds as they use their custom Json AOT.

Changin Json.net from a fixed version to a range from to 10 to 12 (not inclusive)