Websockets streaming (with handlers), Observables, subscriptions, etc
* Implementation of streaming websockets with responses using specific handlers
This allows to have multiple requests of the same type using different topics.
* Websocket streaming it is to be used separate from Web3 similar to Unity, the client when Starts connects and start processing.
* There are few generic classes, the main are:
1. RpcStreamingSubscriptionObservableHandler, generic class that includes Subjects and Observables to subscribe, get the data and unsubscribe from a feed. It also handles the state and message, subscription responses.
2. RpcStreamingRequestResponseObservableHandler, any normal RPC method can be used when Streaming (websockets, ipc) this class and further abstractions, implement a request and handler response, including Subject and Observable access.
The subscritions has been imlemented using as a basis the original subscription rpc classes so the build requests are all in the same place.

Test sample is in the WebSocketsStreamingTest Program
It includes subscription to BlockHeaders, PendingTransactions, Logs and Call methods for GetBalance and BlockNumber, there is also demonstration on how to subscribe, unsubscribe, and check for the results of subscribption id etc.

Ideally the Eventing implementation of @wired010 (which is great) can be refactored to use the same websockets class (when finished) and instead of listening to Events use the same handlers, the Events can the be raised in a similar way to the subject handling, maybe using a base class. Sadly due to the different states of Subscriptions these cannot be part of Web3, so they have been removed.

Observables need to be in a different project, as they need support for a higher version of netstandard (2.0).

Async Enumerables, due the different states, subscriptions, request responses would not work very well for a single connection, (as far as I know, have tried) although they could be implemented when reading the output from the websockets (for later).