Initial support for code generating structs
Added support for struct code generation, but as struct types are not know, the simplest workaround until 0.5.8 is to name the struct type the same as the parameter name. So you will have to name for the time being the parameter names the same as the struct type.

If you have the same struct as an input / ouput parameter multiple types for the same function, it will need to be corrected manually.

Namespaces are the same as DTOs.

* .Net model / code generators
* AbiDeserialisation in Typescript
* Test mock up in Javascript
* TODO: .Net des / conversion
* TODO: Migrate to 0.5.8 when available