LogOrchestrator changes to retry on nulls, and report errors whist executing actions #723
There seems to be an issue that Infura is returning null values, when it should be returning some values when processing logs, this implements a simple retry mechanism when null values are returned, although this adds an small retry overload (single request), it should not be an issue as block ranges are defaulted to 100.

Included is the resetting of retries, and error handling when calling actions, these should be handled already by users, but this does not updated the progress.