#361 #634 #730 #428 V4 Array / Tuple support for typed messages
Enabling simpler input of typed messages on 712 using complex types and arrays (v4). Typed as in c# types.. The simplest pattern (for users I could think of) is to create a generic typedMessage as in 712 that can be reused across for that specific type of message, then using type c# classes to input the message.