MulticallQuery upgrade, common contract addresses as constants, CallInput use data instead of input, ERC721 multicall to retrive all the user tokens
Multicall query support for v3, and as a default now. To call v1 only supported multicall a new method (v1) has been added.
EthCall revert to data instead of input, as this is not supported in some chain forks
CommonContracts addresses as constants.
ERC721 service extensions to retrieve all the tokens of a user, and urls