#872 Clarifying that we will update the last block processed even on Cancellation.
This cannot rely on internal implementations as these might not follow the same process.

The issue on 872 was on ``` fromBlockNumber = progress.BlockNumberProcessTo.Value + 1;``` the cancellation before any progress will mean that it will fail on this assignment as it will be null. Checking for cancellation solves this.

UpdateLastBlockProcessedAsync already validates if the block number is null and in this scenario won't be updated.