Evm simulator with support and addition of rpc request DebugStorageRangeAt to enable debug / simulate existing transactions (needs an archive node that supports it)
DebugStorageRangeAt returns the state at a specific transactionIndex, this is necessary as StorageAt is block based, and the state can have changed in a previous executed transaction when replaying / simulating an existing transaction. Tests modified to match behaviour, but an archive node is required that supports it.
Other general refactorings.