Metamask Unity Webgl Web3 / interceptor / host provider support
Added support for Metamask to work in Unity using Web3 with an interceptor the same as in Blazor.
MetamaskWebGlHostProvider.CreateOrGetCurrentInstance() will create a new instance.
MetamaskTaskRequestInterop provides the IMetamaskInterop implementation to be used with the Host / Web3 for interception, including callbacks for SelectedNetworkChanged and  SelectedAccountChanged which will use the webgl host to raise events synchronously.