New TableRepositoryBase to use across different repository implementations
This includes the capability to query predicates, although not supported InMemory as we can use Linq directly (simpler).. for this a new object is exposed AllRecords.

InMemoryTableRepository stores the keys without any concatenation.

ITableRepository now expects as an input the BlockNumber, address, logIndex  so they can be stored

StoredRecord includes these values and upto 4 keys, all values are set using bytes to improve performance (lazyload to hex) and storage on db