Mud.Repositories.EntityFramework
Project providing the base objects to integrate with a mud repository using the Entity Framework.
It provides a custom predicate builder to create sql queries using hex values using the StoredRecords, the statements are created as a SqlPredicateResult  providing both the sql query and arguments to prevent injection.

A MudEFTableRepository using keys as hex is also provided, which will be used to save onto a context that implements IMudStoreRecordsDbSets

A BlockProgressRepository supporting the Log Processor and Blockchain Processor progress