NethereumMudLogProcessing, NethereumMudStoredRecordsRestApi and NethereumMudStoreRestApiClient
Console tests and examples of an e2e solution in which, store records are processed and stored in a postgres db using the LogProcessor, (including migrations etc using EF), and then queried using a predicate, which is converted into json and pass as a request to a Rest Api which then converts into a Predicate to query the postgres db using the postgres table repository. The result is transformed onto a DTO which is the transformed onto a table and output to the console.