Postgres StoreRecords Normaliser
Processing service to normalise all the tables and records from the StoredRecords table populated by the Store event logs.
Process, for each store record, it validates that the table matching the tableId has been created, if not a new one will be created using the column mapping types from the schema in the blockchain. Per each row all the data will be decoded in a generic format and then inserted or updated on the database.

Singleton tables are always created with an Id and key of 1 to simplify them.
An Progress table is also is created in the database to keep track of the last blocknumber and rowId processed so the processing can be paged (configurable)