Mud InMemoryChangeSet TableRepository trackers
ChangeSet tracker that allows to get all the changes added to the a new InMemoryChangeTrackerTableRepository, eventually this can be moved to the base class if deem necessary as it can be used across.

This allows to sync models with changes that have been processed, like a receipt or combination of logs.